requirements.txt
gunicorn==23.0.0
packaging==24.2
asgiref==3.8.1
cx_Oracle==8.3.0
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
numpy
openpyxl==3.1.5
pandas
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
